move ensure_data_types, and use in computation of incicies
